backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
aiofiles==24.1.0
pandas==2.2.3
numpy==1.26.4
torch==2.5.1
transformers==4.47.1
spacy==3.8.2
scikit-learn==1.5.2
nltk==3.9.1
faiss-cpu==1.9.0
sentence-transformers==3.3.1
groq==0.13.0
networkx==3.4.2
tqdm==4.67.1
